fix(jwt): possible NPE in jwt:encode
refs #2

The implementation of `jwt:encode` with arrow expressions failed with an NPE.
It does so only when called in a test case as
part of an XQSuite.
Since this module is considered security relevant code
the functon was refactored to be extra sure this will work under all
circumstances.

If the underlying NPE of the XQuery runtime is fixed the implementation can use arrow expressions again.